scrapper/forums/requirements.txt
# Web scraping dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71

# LLM
google-generativeai>=0.8.5

# Utilities
python-dotenv>=1.0.0
